Changed main.py slightly and moved ensembl_api.py and gene lists under dna to make the merging compatible with the main branch.
