--- apis/ncbi_api/readme.txt
@@ -1,1 +1,0 @@
-
