--- dataset_creation/rna/readme.txt
@@ -1,1 +1,0 @@
-
